src/main/java/com/artemnizhnyk/junitapp/Dog.java
package com.artemnizhnyk.junitapp;

public class Dog {

    private String name;
    private int age;

    public Dog(final String name, final int age) {
        this.name = name;
        this.age = age;
    }

    public String getName() {
        return name;
    }

    public void setName(final String name) {
        if (this.name.isEmpty()) {
            this.name = name;
        }
    }

    public int getAge() {
        return age;
    }

    public void setAge(final int age) {
        this.age = age;
    }
}
